+++ pkg/ingester/metrics.go
@@ -30,28 +30,29 @@ const (
 )
 
 type ingesterMetrics struct {
-	ingestedSamples         prometheus.Counter
-	ingestedHistograms      prometheus.Counter
-	ingestedExemplars       prometheus.Counter
-	ingestedMetadata        prometheus.Counter
-	ingestedSamplesFail     prometheus.Counter
-	ingestedHistogramsFail  prometheus.Counter
-	ingestedExemplarsFail   prometheus.Counter
-	ingestedMetadataFail    prometheus.Counter
-	oooLabelsTotal          *prometheus.CounterVec
-	queries                 prometheus.Counter
-	queriedSamples          prometheus.Histogram
-	queriedExemplars        prometheus.Histogram
-	queriedSeries           prometheus.Histogram
-	queriedChunks           prometheus.Histogram
-	memSeries               prometheus.Gauge
-	memMetadata             prometheus.Gauge
-	memUsers                prometheus.Gauge
-	memSeriesCreatedTotal   *prometheus.CounterVec
-	memMetadataCreatedTotal *prometheus.CounterVec
-	memSeriesRemovedTotal   *prometheus.CounterVec
-	memMetadataRemovedTotal *prometheus.CounterVec
-	pushErrorsTotal         *prometheus.CounterVec
+	ingestedSamples          prometheus.Counter
+	ingestedHistograms       prometheus.Counter
+	ingestedExemplars        prometheus.Counter
+	ingestedMetadata         prometheus.Counter
+	ingestedSamplesFail      prometheus.Counter
+	ingestedHistogramsFail   prometheus.Counter
+	ingestedExemplarsFail    prometheus.Counter
+	ingestedMetadataFail     prometheus.Counter
+	ingestedHistogramBuckets *prometheus.HistogramVec
+	oooLabelsTotal           *prometheus.CounterVec
+	queries                  prometheus.Counter
+	queriedSamples           prometheus.Histogram
+	queriedExemplars         prometheus.Histogram
+	queriedSeries            prometheus.Histogram
+	queriedChunks            prometheus.Histogram
+	memSeries                prometheus.Gauge
+	memMetadata              prometheus.Gauge
+	memUsers                 prometheus.Gauge
+	memSeriesCreatedTotal    *prometheus.CounterVec
+	memMetadataCreatedTotal  *prometheus.CounterVec
+	memSeriesRemovedTotal    *prometheus.CounterVec
+	memMetadataRemovedTotal  *prometheus.CounterVec
+	pushErrorsTotal          *prometheus.CounterVec
 
 	activeSeriesPerUser        *prometheus.GaugeVec
 	activeNHSeriesPerUser      *prometheus.GaugeVec
@@ -130,6 +131,14 @@ func newIngesterMetrics(r prometheus.Registerer,
 			Name: "cortex_ingester_ingested_metadata_failures_total",
 			Help: "The total number of metadata that errored on ingestion.",
 		}),
+		ingestedHistogramBuckets: promauto.With(r).NewHistogramVec(prometheus.HistogramOpts{
+			Name:                            "cortex_ingester_ingested_histogram_buckets",
+			Help:                            "The number of ingested native histogram buckets per user.",
+			NativeHistogramBucketFactor:     1.1,
+			NativeHistogramMaxBucketNumber:  100,
+			NativeHistogramMinResetDuration: 1,
+			Buckets:                         prometheus.ExponentialBuckets(1, 2, 10), // 1 to 512 buckets
+		}, []string{"user"}),
 		oooLabelsTotal: promauto.With(r).NewCounterVec(prometheus.CounterOpts{
 			Name: "cortex_ingester_out_of_order_labels_total",
 			Help: "The total number of out of order label found per user.",
@@ -359,6 +368,7 @@ func (m *ingesterMetrics) deletePerUserMetrics(userID string) {
 	m.usagePerLabelSet.DeletePartialMatch(prometheus.Labels{"user": userID})
 	m.limitsPerLabelSet.DeletePartialMatch(prometheus.Labels{"user": userID})
 	m.pushErrorsTotal.DeletePartialMatch(prometheus.Labels{"user": userID})
+	m.ingestedHistogramBuckets.DeleteLabelValues(userID)
 
 	if m.memSeriesCreatedTotal != nil {
 		m.memSeriesCreatedTotal.DeleteLabelValues(userID)
@@ -621,7 +631,7 @@ func newTSDBMetrics(r prometheus.Registerer) *tsdbMetrics {
 		tsdbOOOHistogram: prometheus.NewDesc(
 			"cortex_ingester_tsdb_sample_ooo_delta",
 			"Delta in seconds by which a sample is considered out of order (reported regardless of OOO time window and whether sample is accepted or not).",
-			nil, nil),
+			[]string{"user"}, nil),
 		tsdbMmapChunksTotal: prometheus.NewDesc(
 			"cortex_ingester_tsdb_mmap_chunks_total",
 			"Total number of chunks that were memory-mapped.",
@@ -764,7 +774,7 @@ func (sm *tsdbMetrics) Collect(out chan<- prometheus.Metric) {
 	data.SendSumOfCountersPerUserWithLabels(out, sm.tsdbOOOSamples, "prometheus_tsdb_out_of_order_samples_total", "type")
 	data.SendSumOfCountersPerUserWithLabels(out, sm.tsdbOutOfOrderSamplesAppended, "prometheus_tsdb_head_out_of_order_samples_appended_total", "type")
 	data.SendSumOfCounters(out, sm.tsdbSnapshotReplayErrorTotal, "prometheus_tsdb_snapshot_replay_error_total")
-	data.SendSumOfHistograms(out, sm.tsdbOOOHistogram, "prometheus_tsdb_sample_ooo_delta")
+	data.SendSumOfHistogramsPerUser(out, sm.tsdbOOOHistogram, "prometheus_tsdb_sample_ooo_delta")
 	data.SendSumOfCounters(out, sm.tsdbMmapChunksTotal, "prometheus_tsdb_mmap_chunks_total")
 	data.SendSumOfCounters(out, sm.checkpointDeleteFail, "prometheus_tsdb_checkpoint_deletions_failed_total")
 	data.SendSumOfCounters(out, sm.checkpointDeleteTotal, "prometheus_tsdb_checkpoint_deletions_total")